Move classes to separate files, configure tools by decorator
